[es7] remove async/await (#368)
# Description
Babel.js is not able to handle the async/await feature without a polyfill to be included along with the SDK on production.
Forcing the use of a polyfill is not a desired outcome for our SDK, so async/await is now a banned feature.

# How-to test
The "raw web" getting started tests are currently failing on the documentation because of this problem. Using this branch version solve the failing test.