Prevent pending request leak when disconnect the SDK (#424)
## What does this PR do?

When a request is made to Kuzzle, we setup a once listener to handle the response. If a disconnection occur (manual with `disconnect()` or not), the listener are never freed.  

This PR save the pending requests in a `Map` and when a disconnection occur, a `discarded` event is triggered for each requests and then the map is cleared.

Fix #356 

### Other changes

  - add .npmignore