Throw an error when an URL param is missing in HTTP (#490)
## What does this PR do?

When you send a request with arguments that are present in the URL (eg: `/:index/_create`), they are injected in the URL.  
If the user didn't put the required arguments, they are `undefined` and they end up in the url as the string `"undefined"` (eg: `/undefined/create`).    

This PR now discard the request if an URL param is missing.