Hotfix document search http (#498)
## What does this PR do?

When we have multiple choose for a route, we choose the shortest or the GET route.  
With `document:create` we have the choice between those 2 routes:
```
{ verb: 'GET', url: '/:index/:collection/' },
{ verb: 'POST', url: '/:index/:collection/_search' }
```

The problem is that after even if we want to change the verb by using `options.verb`, the URL is not the good one and `GET /:index/:collection` correspond to the `collection:create` method.

So this PR is just a dirty fix to always use the POST route with `document:search`.  

When we are going to fully support the GET verb with query for `document:search`, maybe we can get ride of this.

See this failing job on Kuzzle: https://travis-ci.org/github/kuzzleio/kuzzle/jobs/667658863