Fix re-subscription (#602)
When resubscribing after a reconnected event, the SDK use the original subscription request to re-subscribe.

The Kuzzle.query method mutate request and inject the JWT inside. Meaning that the request saved for the re-subscription already has a JWT inside.

This can be an issue because developers rely on the kuzzle.jwt property of the SDK to know if 1) the token is still valid 2) there are connection information.

Now the request and options parameter of the query method are cloned before modification.