[HOTFIX] Ensure returning the user body in the content property when serialized (#630)
Since the last release, the user content is accessible under the `user._source` property in addition to the `user.content` property.

A getter was added for backward compatibility but when a User object was returned in a controller action it was serialized with only the `_source` property, breaking existing APIs relying on the `content` property.

This PR fix that but the sent data will be duplicated. A safe `serialize` method has been added to the User, Role and Profile classes and should be used instead of returning the instances directly.