fix: aborted requests not rejected (#644)
When requests are queued, there are 2 constructor options that can lead to abort them before they can be sent to Kuzzle:

queueTTL: if set to a value larger than 0, queued requests can expire
queueMaxSize: if set to a value larger than 0, the queue has a maximum number of requests it can hold. If reached, older requests are aborted to make place for newer ones
When requests are aborted because of those configurations, an offlineQueuePop event is emitted, but:

the invalid argument is provided, making the event deliver an undefined payload
the promise associated to the aborted request is not rejected, making clients await indefinitely for them
This PR fixes these issues, and add the missing unit tests that should have caught these problems earlier.

Other changes
the offlineQueuePop event emitted when the queue is being played does not deliver the right payload. The entire queued object is delivered (containing the queued request and additional information about it), instead of just the request, as per the documentation. So, I fixed that too.