Add authenticator function used at reconnection (#650)
When the SDK reconnect to Kuzzle, it trigger the `reconnected` event. The Realtime controller will try to resubscribe when this event is triggered.

If the token had expired, then the Realtime controller will try to resubscribe with no authentication and thus the subscriptions request may fail.

This PR include a new `authenticator` property, this property should contain a function that authenticate the SDK (with `auth.login` for example).  
The SDK will call the function before emitting the reconnected event, if the SDK was authenticated and cannot re-authenticate then the `reconnected` event will not be emitted and the SDK will be in the `disconnected` state.

A new `reconnectionError` has been added and is triggered when the reconnection has failed