Allow to use `refresh: false` with `document:upsert|deleteByQuery` (#717)
As mentioned in #700 , the refresh option type was not well defined in this section. This as been fixed by @Aschen in the past month but when checking the code before closing the issue, I noticed there were still two action of the document controller that still have a wrongly defined type for this field. In order to close #700, this PR aim to update these